Refactor folder structure into leaderboard entity
