--- apps/veil/app/api/position/leaderboard/route.ts
@@ -1,1 +1,0 @@
-export { GET } from '@/pages/leaderboard/server';
