--- apps/veil/src/pages/leaderboard/server.ts
@@ -1,1 +1,0 @@
-export { GET } from '@/entities/leaderboard/server';
